Notebook | Keyword extraction and Coursera dataset
In this notebook, I first tried to create the simple steps of how to clean and extract the keywords from two different PDFs and try to match it (recommendation feature). 
The first is that one about the Big Bang, that we mentioned on our last meeting. It is a smaller article that contains some pictures and that returns, as one of its keywords, an expression ("Big Bang"), instead of only single words. This is an important point to address, when we start shaping the code.
The second is an article about Finance. Much longer and much more "old style". I wanted to check if everything would work.

Then, I manually created a few courses, in order to try a first match between the keywords extracted earlier and the ones from the courses.

After that, I found a Coursera Dataset on Kaggle and did some cleaning. I am still working on it, but I decided to share with you guys, so you can take a look and we can exchange some thoughts for the next steps.
Here is the link to the dataset> https://www.kaggle.com/datasets/elvinrustam/coursera-dataset